Fixup the circular buffer filling.
Firstly, invalidate the buffer by setting both indicies to be invalid: this prevents the deliver kernel from delivering events from this neuron
(this might fix some lockups on the GPU related to this).

Also, the old code didn't handle the case when the buffer end looped correctly (possible invalid write, but probably not something that ever
happened in practice).